// src/services/types.ts
export interface RootAPIResponse<T> {
  hasError: boolean
  value: T
  error: Error
}

export interface PaginatedResult<T> {
  currentPage: number
  pageCount: number
  pageSize: number
  totalRecords: number
  result: T[]
}

export interface AdminUser {
  firstName: string
  lastName: string
  emailAddress: string
}

export interface Error {
  errorCode: string
  message: string
  isTransient: boolean
  inner: string
  fullMessage: string
}
// ...
export interface IConfigurationEntry {
  key: string
  value: string
  isSecret: boolean
  isNew?: boolean
}

export interface ISetting {
  settingId?: string
  provider?: string
  configurationEntries: IConfigurationEntry[]
}